enabled 3d AA, changed default script for script panel, small testing changes
Tested and discovered that 2d AA breaks the 3d scene rendering with openxr, also changed the default script that is autofilled by the scripting panel so it can instantly reflect changes and have some basic useful template code.

When testing the scripting panel, you will likely have to disable warnings or run the project though the project selection dialog because the function that parses the script throws a proper error status which the debugger catches making it difficult to test with.